Refactor for seperation and copilot suggestions
